third_party/blink/web_tests/virtual/document-pip/document-pip-returns-window-with-document.html

<!DOCTYPE html>
<title>Test that window.requestPictureInPictureWindow() returns a
  PictureInPictureWindow object with a document</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
  <div id="div"></div>
<script>
const div = document.getElementById('div');

async_test((t) => {
  window.requestPictureInPictureWindow().then(t.step_func_done(
      (pipWindow) => {
    assert_true(!!pipWindow.document,
        'We should receive a PictureInPictureWindow with a document');

    // We should be able to move an element to that document.
    assert_true(document.body.contains(div),
        'The original document should start with the div');
    assert_false(pipWindow.document.body.contains(div),
        'The PiP document should not start with the div');
    pipWindow.document.body.append(div);
    assert_false(document.body.contains(div),
        'The div should have moved away from the original document');
    assert_true(pipWindow.document.body.contains(div),
        'The div should have moved to the PiP document');
  }));
});
</script>
